+++ Parser.h
@@ -2,4 +2,4 @@
 
 network read_network(string);
 vector<vector<string> > readData(int, string);
-void out(string, vector<pair<string, vector<float> > >);
+void out(string, vector<pair<string, vector<double> > >);
